Extract createSubquery in PostgresTLPBase and PostgresExpressionGenerator into new PostgresUtils
