+++ src/sqlancer/postgres/oracle/tlp/PostgresTLPBase.java
@@ -1,5 +1,7 @@
 package sqlancer.postgres.oracle.tlp;
 
+import static sqlancer.postgres.PostgresUtils.createSubquery;
+
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -19,7 +21,6 @@ import sqlancer.postgres.PostgresSchema.PostgresDataType;
 import sqlancer.postgres.PostgresSchema.PostgresTable;
 import sqlancer.postgres.PostgresSchema.PostgresTables;
 import sqlancer.postgres.ast.PostgresColumnValue;
-import sqlancer.postgres.ast.PostgresConstant;
 import sqlancer.postgres.ast.PostgresExpression;
 import sqlancer.postgres.ast.PostgresJoin;
 import sqlancer.postgres.ast.PostgresSelect;
@@ -67,8 +68,7 @@ public class PostgresTLPBase extends TernaryLogicPartitioningOracleBase<Postgres
         // JOIN subqueries
         for (int i = 0; i < Randomly.smallNumber(); i++) {
             PostgresTables subqueryTables = globalState.getSchema().getRandomTableNonEmptyTables();
-            PostgresSubquery subquery = PostgresTLPBase.createSubquery(globalState, String.format("sub%d", i),
-                    subqueryTables);
+            PostgresSubquery subquery = createSubquery(globalState, String.format("sub%d", i), subqueryTables);
             PostgresExpression joinClause = gen.generateExpression(PostgresDataType.BOOLEAN);
             JoinType options = JoinType.getRandomForDatabase("POSTGRES");
             PostgresJoin j = new PostgresJoin(subquery, joinClause, options);
@@ -108,35 +108,6 @@ public class PostgresTLPBase extends TernaryLogicPartitioningOracleBase<Postgres
     @Override
     protected ExpressionGenerator<PostgresExpression> getGen() {
         return gen;
-    }
-
-    public static PostgresSubquery createSubquery(PostgresGlobalState globalState, String name, PostgresTables tables) {
-        List<PostgresExpression> columns = new ArrayList<>();
-        PostgresExpressionGenerator gen = new PostgresExpressionGenerator(globalState).setColumns(tables.getColumns());
-        for (int i = 0; i < Randomly.smallNumber() + 1; i++) {
-            columns.add(gen.generateExpression(0));
-        }
-        PostgresSelect select = new PostgresSelect();
-        select.setFromList(tables.getTables().stream().map(t -> new PostgresFromTable(t, Randomly.getBoolean()))
-                .collect(Collectors.toList()));
-        select.setFetchColumns(columns);
-        if (Randomly.getBoolean()) {
-            select.setWhereClause(gen.generateExpression(0, PostgresDataType.BOOLEAN));
-        }
-        if (Randomly.getBooleanWithRatherLowProbability()) {
-            select.setOrderByClauses(gen.generateOrderBys());
-        }
-        if (Randomly.getBoolean()) {
-            select.setLimitClause(PostgresConstant.createIntConstant(Randomly.getPositiveOrZeroNonCachedInteger()));
-            if (Randomly.getBoolean()) {
-                select.setOffsetClause(
-                        PostgresConstant.createIntConstant(Randomly.getPositiveOrZeroNonCachedInteger()));
-            }
-        }
-        if (Randomly.getBooleanWithRatherLowProbability()) {
-            select.setForClause(ForClause.getRandom());
-        }
-        return new PostgresSubquery(select, name);
     }
 
 }
